10 - Lock-On mechanic
Implimented the Lock-on mechanis with the ability to walk, run and dodge around the closest enemy to the player. The camera will change to an over the shoulder look and return to the free-look when unlocked.
A little clean up of of Character Movement scripts Update block for better readability, further clean up of the code is required.