12 - Freelook aiming, anim clear up and aditional actions of lockon
Made the player character look in the direction of the camera when firing in free look mode when not moving and having an arrow drawn.
Made the anim controller more clean putting dodges, firing and additional actions in their own sub-states.
Added the ability to punch and kick while in lockon mode.