13 - Falling
Implimented the falling anims for the player, when a boxcast no longer picks up any colisions a falling loop will start. If there is any input when landing a smoother anim for going into walking or running will trigger. if there is no input uppon landing then a better anim for going into idle is played.