15.3 - Arrows and visual upgrade
Made the firing action spawn an arrow that then as a forward force applied. currently collides with object, other than terrains, but doesn't perform any mechanics
Made a prefab of the bow unattached to the archer for a back model, this model will be visual on the back of the character when the bow is disarmed.
Made a arrow prefab for the projectile used for shooting but also for the drawing and firing an arrow animation, the arrow is attached to the pointing finger, becoming visable when an arrow is drawn and held. Disapparing when fired.