20.1 - Start of re-design
Due to clunkyness and some bugs due to anims and anim events. I have decided to redesign the animator controller structure and make use of state behaviuour scripts.
I will go on to change the Audio as well to make it easier to impliment, change and fine tune using scriptable objects.