23.1 - Footstep Bugfix
The bug with footstep SFX continuing to play when holding down any movement input as been fixed through use of StateMachineBehavious script that turns a bool true and false when entered and exited. This allows the audio manager to know if the player is currently within a marked sub-state aka an action meaning the footstep sound effect will not play.